[Refactor/#135] CustomerCenterScreen을 screens.customer_center 패키지로 이동
